[feature/grade-cal] 이수학점 및 성적 현황 한 화면에 통합
