+++ src/MainShell.js
@@ -7,7 +7,7 @@ import Main from './Main';
 import Menu from './Menu';
 import AttendanceScreen from './screens/attendance';
 import Schedules from './screens/schedules';
-import Credits from './screens/credits';
+import {CreditsAndGrade, Credits, GradeChart} from './screens/creditsAndGrade';
 import StudentTimetable from './screens/studentTimetable';
 import CounselHistory from './screens/counsel';
 import Syllabus from './screens/syllabus';
@@ -20,7 +20,6 @@ import About from './screens/about';
 import {Meal} from './screens/meal';
 import searchCondition from './screens/searchCondition';
 import {MaterialIcons} from '@expo/vector-icons';
-import BuildConfigs from './config';
 import {ProfessorTimetable, SearchProfessors} from './screens/professorTimetable';
 import {LectureRoomTimetable, SearchLectureRooms} from './screens/lectureRoomTimetable';
 import Authinfo from './screens/authinfo';
@@ -45,11 +44,11 @@ function HomeStack(){
       <HStack.Screen name="Home" component={Main} options={{headerShown: false}}/>
       <HStack.Screen name="Attendance" component={AttendanceScreen} options={{title: '나의 출결 현황'}}/>
       <HStack.Screen name="Schedules" component={Schedules} options={{title: '학사 일정'}}/>
-      <HStack.Screen name="Credits" component={Credits} options={{title: '현재 이수 학점'}}/>
       <HStack.Screen name="Timetable" component={StudentTimetable} options={{title: '시간표'}}/>
       <HStack.Screen name="SyllabusDetails" component={SyllabusDetails} options={{title: '강의계획서 상세'}}/>
       <HStack.Screen name="Meal" component={Meal} options={{title: '주간 식단'}}/>
       <HStack.Screen name="Authinfo" component={Authinfo} options={{headerShown: false}}/>
+      <HStack.Screen name="CreditsAndGrades" component={CreditsAndGrade} options={{title: '학점/성적'}}/>
     </HStack.Navigator>
   );
 }
