+++ src/screens/creditsAndGrade.js
@@ -1,12 +1,93 @@
-import React, {useEffect, useState} from 'react';
+import React, {Component, useState, useEffect} from 'react';
+import {View, FlatList, Dimensions, ActivityIndicator} from 'react-native';
 import ListItem from '../components/listitem';
-import {View, FlatList, ActivityIndicator, Dimensions} from 'react-native';
 import ForestApi from '../tools/apis';
 import BuildConfigs from '../config';
-import {useTheme} from '@react-navigation/native';
 import {ThemedText, Collapsible} from '../components/components';
-import { LineChart} from "react-native-chart-kit";
-export default function GradeChart(props){
+import {LineChart} from "react-native-chart-kit";
+import {useTheme} from '@react-navigation/native';
+
+import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
+
+const Tab = createMaterialTopTabNavigator();
+
+export function CreditsAndGrade() {
+  return (
+    <Tab.Navigator>
+      <Tab.Screen name="Credits" component={Credits} options={{title: "이수 학점"}}/>
+      <Tab.Screen name="GradeChart" component={GradeChart} options={{title: '성적 현황'}}/>
+    </Tab.Navigator>
+  );
+}
+
+
+export class Credits extends Component{
+  constructor(props){
+    super(props);
+    this.state = {
+      data: [],
+      summary: '',
+      isLoading: false
+    };
+  }
+  async componentDidMount(){
+    this.setState({isLoading: true});
+    const credits = await ForestApi.get('/user/credits', true);
+    if(credits.ok){
+      let finalArr = [];
+      let data = await credits.json();
+      const chunk = 3;
+      for (let i=0, j=data.credits.length; i<j; i+=chunk) {
+        let tempArr = data.credits.slice(i, i+chunk);
+        finalArr.push(tempArr);
+      }
+      this.setState({
+        data: finalArr,
+        summary: data.summary,
+        isLoading: false
+      });
+    }
+  }
+  render(){
+    if(this.state.isLoading){
+      return(
+        <View style={{justifyContent: 'center', padding: 32}}>
+          <ActivityIndicator size="large" color={BuildConfigs.primaryColor} />
+        </View>
+      );
+    }else{
+      return(
+        <View>
+          <FlatList style={{height: '100%'}}
+            data={this.state.data}
+            keyExtractor={(item, index) => index}
+            ListFooterComponent={()=>(
+              <ListItem>
+                <ThemedText style={{fontWeight: 'bold', textAlign: 'center'}}>
+                  {this.state.summary}
+                </ThemedText>
+              </ListItem>
+            )}
+            renderItem={({item})=>
+              <ListItem style={{flex: 1, flexDirection: 'row'}}>
+                {item.map((subItem, index)=>{
+                  return(
+                    <View style={{flex: 1}}>
+                      <ThemedText style={{fontWeight: 'bold',  textAlign: 'center'}}>{subItem.type}</ThemedText>
+                      <ThemedText style={{textAlign: 'center'}}>{subItem.earned}</ThemedText>
+                    </View>
+                  );
+                })}
+              </ListItem>
+            }/>
+        </View>
+      );
+    }
+      
+  }
+}
+
+export function GradeChart(props){
   const {colors} = useTheme();
   const [details, setDetails] = useState([]);
   const [isLoading, setIsLoading] = useState(false);
